Revert "Revert "From Disk To RAM  Reading Hopefuly""
This reverts commit 8939ca64b53867d7a2344c19fba1d996c24fd2fa.

+++ Makee/Makee/Variables.vb
@@ -13,6 +13,6 @@ Public Class Variables
     Public Shared InMachines As Boolean
     Public Shared Achievements As Boolean
     Public Shared ChunksDirectory() As String
-    Public Shared ChunksValues(,,) As Byte
+    Public Shared ChunksValues(,,) As UShort
     Public Shared R As New Random
 End Class
